Drawing quad and images

+++ Source Code/Vec2.h
@@ -5,12 +5,12 @@ struct Vec2
 {
 	//Methods
 	Vec2() {};
-	Vec2(int x, int y) : x(x), y(y) {};
+	Vec2(float x, float y) : x(x), y(y) {};
 
-	void Set(int x, int y) { this->x = x; this->y = y; };
+	void Set(float x, float y) { this->x = x; this->y = y; };
 
 	//Variables
-	int x = 0, y = 0;
+	float x = 0, y = 0;
 };
 
 #endif
